+++ src/pages/donate-page/RightSection/RightSection.tsx
@@ -1,8 +1,8 @@
 import React from 'react';
 import './RightSection.scss';
 import { IN_UKRAINE_LABEL, NOT_IN_UKRAINE_LABEL } from '../../../const/donate-page/donate-page';
-import { UkrainePaymentDetails } from './UkrainePaymentDetails/UkrainePaymentDetails';
-import { AbroadPaymentDetails } from './AbroadPaymentDetails/AbroadPaymentDetails';
+import { UkrainePaymentDetails } from './Ukraine-payment-details/UkrainePaymentDetails';
+import { AbroadPaymentDetails } from './abroad-payment-details/AbroadPaymentDetails';
 
 export const RightSection = () => {
     const [isAbroad, setIsAbroad] = React.useState(false);
