+++ src/pages/donate-page/copy-text-button/CopyTextButton.scss
@@ -1,0 +1,5 @@
+.copyTextButton {
+    cursor: pointer;
+    width: 1.25rem;
+    height: 1.25rem;
+}
